Separate common test header for stream and the rest of the core
