+++ test/core/stream/filter1.cpp
@@ -13,7 +13,7 @@
 #define BOOST_TEST_DYN_LINK
 #define BOOST_TEST_MODULE filter1_test
 
-#include "../common.hpp"
+#include "common.hpp"
 
 #include <hipipe/core/stream/filter.hpp>
 #include <hipipe/core/stream/for_each.hpp>
