+++ test/core/stream/pad.cpp
@@ -11,7 +11,7 @@
 #define BOOST_TEST_DYN_LINK
 #define BOOST_TEST_MODULE stream_pad_test
 
-#include "../common.hpp"
+#include "common.hpp"
 
 #include <hipipe/core/stream/pad.hpp>
